Added tests for genMuXZSd, genLatentData, getLassoLambda, and getModelSize; small tweaks to code based on tests
genLatentData had been assigning arguments to functions in the wrong order internally (didn't cause issues for simulations in paper, but caused issues for other settings). getModelSize didn't handle inputs robustly enough to be user-facing (originally I intended getModelSize to only be internal, but it is a useful function for end-users, and it wasn't difficult to modify it to be user-friendly). Also updated documentation for getModelSize to be user-facing.