Added documentation about issue with clusters argument; added tests for protolasso and clusterRepLasso
Currently debugging issues with beta output in protolasso tests, as well as issues in tests for clusterRepLasso.

Note added to documentation about clusters argument: in testing I noticed that if you provide a data.frame that contains a categorical variable with more than 2 categories, this of course leads to model.matrix creating a matrix with more columns than the provided data.frame. The issue is that the code doesn't account for this when handling the provided clusters argument, so the resulting cluster assignments don't match what the user likely intended. For now, I added a TODO about this and also added a note about this to the clusters argument in all exported functions.